attendance/backend/requirements.txt
fastapi
uvicorn
face-recognition
opencv-python-headless
pillow
numpy
